requirements.txt
Flask==3.1.0
Flask_Cors==5.0.0
Flask_JWT_Extended==4.7.1
passlib==1.7.4
PyJWT==2.10.1
PyJWT==2.10.1
pymongo==4.12.0
python-dotenv==1.1.0
